Modificado editar atracciones ya conserva la imagen si no se va a modificar

+++ src/app/crear/crear.component.ts
@@ -13,7 +13,8 @@ import swal from "sweetalert2";
 export class CrearComponent {
   lugar: any = {};
   id: any = null;
-  file: any;
+  file: any = null;
+  newFile: Boolean = false;
   constructor(
     private lugaresService: LugaresService,
     private route: ActivatedRoute,
@@ -39,16 +40,18 @@ export class CrearComponent {
       this.lugar.lng = result.results[0].geometry.location.lng;
       if (this.id != "new") {
         this.lugaresService.editarLugar(this.lugar);
-        const filePath = "atracciones/" + this.lugar.id;
-        this.storage.ref(filePath).delete();
-        const fileRef = this.storage.ref(filePath);
-        const task = this.storage.upload(filePath, this.file);
+        if(this.newFile == true && this.file != null){
+          const filePath = "atracciones/" + this.lugar.id;
+          this.storage.ref(filePath).delete();
+          const fileRef = this.storage.ref(filePath);
+          const task = this.storage.upload(filePath, this.file);
+        } 
         swal.fire(
           "Atraccion Editada",
           "Tu atraccion se a editado con exito",
           "success"
         );
-        this.router.navigate([`/detalle/${this.id}`]);
+        this.router.navigate([`/usuario`]);
       } else {
         this.lugar.id = Date.now();
         this.lugar.propietario = this.angularFireAuth.auth.currentUser.uid;
@@ -72,7 +75,9 @@ export class CrearComponent {
 
   uploadFile(event) {
     this.file = event.target.files[0];
+    this.newFile = true;
   }
+
   public botFiltro(){
     var grocerias = ["puta","puto","marica","mierda","chingadera","chinga","puteria","mamada","chupala","alv","chupala","verga","pendejo","chingar","mamar","mamando","puteria","chingado","culo","culero","estupido","idiota","baboso","cabron","pito","tarado","tonto","cagas","joto","prostituta","golfa","malparida","malparido","ano","pene","vagina","tetas","chichis","bubis","jodido","madrazo","castra","pinche","emputado","encabronado","bastardo"];
